Fix a printf format type warning in macos's CD-ROM API source
Fixes:

SDL_syscdrom.c:498:45: warning: values of type 'OSStatus' should not be
used as format arguments:

   SDL_SetError ("PBUnmountVol returned %d", err);
                                        ~~   ^~~
                                       %d   (int)

1 warning generated.

+++ src/libs/sdlcd/macosx/SDL_syscdrom.c
@@ -495,7 +495,7 @@ static int SDL_SYS_CDEject(SDL_CD *cdrom)
 
 	if (err != noErr) {
         Unlock ();
-		SDL_SetError ("PBUnmountVol returned %d", err);
+		SDL_SetError ("PBUnmountVol returned %d", (int)err);
 		return -4;
 	}
